// Dolphin-main-main/apps/client/src/types/user.types.ts
export interface CompanyInfo {
  businessName: string;
  contactPerson: string;
  profilePicture?: string;
  companyAddress: string;
  companyContactNumber: string;
  pincode: string;
  POCEmailVerified: boolean;
  POCPhoneVerified: boolean;
  state: string;
  city: string;
  contactNumber: string;
  contactEmail: string;
  brandName: string;
  companyEmail: string;
  companyLogoUrl?: string;
  website?: string;
}

export type BusinessType = "b2b" | "b2c" | "d2c";

export interface IUserProfileDB {
  id: string;
  userId: string;
  name?: string;
  email?: string;

  onboardingStep: number;
  monthlyOrderCount: string;
  onboardingComplete: boolean;
  profileComplete: boolean;
  salesChannels: Record<string, boolean>;

  companyInfo: CompanyInfo;
  domesticKyc: DomesticKyc | null;
  bankDetails: BankDetails | null;
  gstDetails: GstDetails | null;
  businessType: BusinessType[];

  approved: boolean;
  approvedAt: string | null;
  rejectionReason: string | null;
  currentPlanId?: string | null;
  currentPlanName?: string | null;

  submittedAt: string;
  updatedAt: string;
}

export type KycStatus =
  | "pending"
  | "verification_in_progress"
  | "verified"
  | "rejected";

export interface DomesticKyc {
  updatedAt: Date | null;
  status: KycStatus;
}

export interface BankAccount {
  accountHolder: string;
  accountNumber: string;
  ifsc: string;
  bankName: string;
  branch?: string;
  id: string;
  accountType?: "Savings" | "Current";
  upiId?: string;
  chequeImageUrl?: string;
  status: "verified" | "pending" | "rejected"; // ✅ verified/unverified status
  rejectionReason?: string;
  isPrimary: boolean;
}

export interface BankDetails {
  count: number; // how many accounts exist
  primaryAccount: BankAccount | null;
}

export interface GstDetails {
  gstNumber: string; // 15‑char GSTIN
  legalName?: string;
  registrationDate?: string; // ISO date
  state?: string;
  documentUrl?: string; // GST cert PDF/image
}